feat(new-issuer): pass downloadOwnershipList as prop to presenter
Also renamed Checkpoints to CheckpointList and Checkpoint to DividendList
