--- packages/new-polymath-issuer/src/components/Checkpoint/index.tsx
@@ -1,3 +1,0 @@
-import { CheckpointContainer } from './Container';
-
-export const Checkpoint = CheckpointContainer;
