+++ packages/new-polymath-issuer/src/components/CheckpointList/index.tsx
@@ -1,0 +1,3 @@
+import { CheckpointListContainer } from './Container';
+
+export const CheckpointList = CheckpointListContainer;
