--- packages/new-polymath-issuer/src/components/Checkpoints/index.tsx
@@ -1,3 +1,0 @@
-import { CheckpointsContainer } from './Container';
-
-export const Checkpoints = CheckpointsContainer;
