+++ packages/new-polymath-issuer/src/components/DividendList/index.tsx
@@ -1,0 +1,3 @@
+import { DividendListContainer } from './Container';
+
+export const DividendList = DividendListContainer;
