+++ packages/new-polymath-issuer/src/components/index.ts
@@ -1,4 +1,4 @@
-export { Checkpoints } from './Checkpoints';
+export { CheckpointList } from './CheckpointList';
 export { DataFetcher, PageLoader } from './enhancers';
 export { ModalTransactionQueue } from './ModalTransactionQueue';
 export { Header } from './Header';
